Add minimal MRIcroS + electrode selection
Bare-bones MRIcroS viewer + file loader, extended with
GeometryController and tweaked to enable electrode selection.
Notably, rotation is implemented by hand instead of using the
built-in rotate3d to bind conditional handlers to the ButtonDown
and ButtonUpFcns needed for identifying and selecting electrodes.

Electrode selection on [PT001]/faceless_postopCT.nii is accurate
with positive inputs. Testing to come for rejecting clicks in
free space. Electrodes are marked by spheres that cover the electrode
surface.