Implement float framework for GUI & begin linkages
Annoyingly, MATLAB `Position` is relative to the bottom-left, which
plays poorly with top-aligned features (e.g. the util panel). There
also isn't much built-in to help mix absolute and relative positioning
(as if everything floating weren't already enough of a pain...)

`FloatingControl` brings the positioning of GUI elements under better
control by addressing those two points. Elements can be aligned to any
edge and both offset and area (width-height) can be mixed- absolute and
relative to the dimensions of the parent container. The elements are
nested in the constructor to fill the gap left behind by a good markup
scheme.

---

Markers are managed by `GridController` and all references to the
current grid and electrode point to the values of the respective
dropdown elements.

Implementation has begun on linkages: visual lines connecting adjacent
selected markers.