Refactor and stabilize, moving MRIcroS into app
MRIcroS has been dismantled and migrated into the application code.
This allows for very aggressive refactoring of the class structure
to separate responsibilities properly and be readable.

The base-level functionality is stable as of this commit. There ought
to be lingering migration bugs around, which will slowly be eeked out
and destroyed with further scrutiny. Another feature also has its
beginnings here: volumetric weighted averaging for locating the centers
of electrodes. Not working yet, but it's probably just the math that's
not quite there yet.